\name{KEGG.db}
\alias{KEGG.db}
\alias{KEGG}
\title{Bioconductor annotation data package}
\description{

  Welcome to the KEGG.db annotation Package.  The purpose of this package
  is to provide detailed information about the latest version of the
  KEGG pathway databases.  This package is updated biannually.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:KEGG.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}
\examples{
  ls("package:KEGG.db")
}
\keyword{datasets}

