\name{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@}
\title{Bioconductor annotation data package}
\description{

  Welcome to the @ANNOBJPREFIX@.db annotation Package.  The purpose of
  this package was to provide detailed information about the latest
  version of the KEGG pathway databases.  But a number of years ago,
  KEGG changed their policy about sharing their data and so this package
  is no longer allowed to be current.  Users who are interested in a
  more current pathway data are encouraged to look at the KEGGREST or
  reactome.db packages.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:@ANNOBJPREFIX@.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}
\examples{
  ls("package:@ANNOBJPREFIX@.db")
}
\keyword{datasets}

